+++ Addons/Metadata/Builtin/Runtime/KnotLocalizedTextOverrideMetadata.cs
@@ -10,38 +10,37 @@ namespace Knot.Localization.Data
 {
     [Serializable]
     [KnotMetadataInfo("Localized Text Override", KnotMetadataInfoAttribute.MetadataScope.Text, AllowMultipleInstances = false)]
-    public class KnotLocalizedTextOverrideMetadata : IKnotTextFormatterMetadata, IKnotCultureSpecificMetadata
+    public class KnotLocalizedTextOverrideMetadata : IKnotTextFormatterMetadata
     {
         public List<LocalizedTextEntry> LocalizedTexts => _localizedTexts ?? (_localizedTexts = new List<LocalizedTextEntry>());
         [SerializeField] private List<LocalizedTextEntry> _localizedTexts = new List<LocalizedTextEntry>();
-
-        [NonSerialized] private CultureInfo _activeCulture;
-
+        
 
         public KnotLocalizedTextOverrideMetadata() { }
 
-        public KnotLocalizedTextOverrideMetadata(params KeyValuePair<string, string>[] customLocalizedTexts)
+        public KnotLocalizedTextOverrideMetadata(params LocalizedTextEntry[] entries)
         {
-            
+            LocalizedTexts.AddRange(entries);
         }
 
         public object Clone() => (MemberwiseClone() as KnotLocalizedTextOverrideMetadata) ?? new KnotLocalizedTextOverrideMetadata();
         
         
-        public void SetCulture(CultureInfo cultureInfo)
-        {
-            _activeCulture = cultureInfo;
-        }
-
         public void Format(StringBuilder sb)
         {
-            var targetEntry = LocalizedTexts.FirstOrDefault(c => CultureInfo.GetCultureInfo(c.CultureName).Equals(_activeCulture));
+            Format(sb, CultureInfo.GetCultureInfo(LocalizedTexts.FirstOrDefault()?.CultureName ?? CultureInfo.DefaultThreadCurrentCulture.Name));
+        }
+
+        public void Format(StringBuilder sb, CultureInfo cultureInfo)
+        {
+            var targetEntry = LocalizedTexts.FirstOrDefault(c => CultureInfo.GetCultureInfo(c.CultureName).Equals(cultureInfo));
             if (targetEntry != null)
             {
                 sb.Clear();
                 sb.Append(targetEntry.Text);
             }
         }
+
 
         [Serializable]
         public class LocalizedTextEntry