+++ Samples~/Demo/Scripts/Metadata/KnotCultureSpecificDayPostfixMetadata.cs
@@ -9,22 +9,17 @@ namespace Knot.Localization.Demo
 {
     [KnotMetadataInfo("Postfix day of the week", KnotMetadataInfoAttribute.MetadataScope.Text)]
     [Serializable]
-    public class KnotCultureSpecificDayPostfixMetadata : IKnotCultureSpecificMetadata, IKnotTextFormatterMetadata
+    public class KnotCultureSpecificDayPostfixMetadata : IKnotTextFormatterMetadata
     {
-        public object Clone() => new KnotCultureSpecificDayPostfixMetadata();
-
-        [NonSerialized]
-        private CultureInfo _currentCulture = CultureInfo.InvariantCulture;
-
         //bug: Unity refuses to serialize classes with no serializable fields, so we put a placeholder to fix it (Issue ID: 1183547)
         [SerializeField, HideInInspector] private bool _serializationPlaceholder;
 
+        public object Clone() => new KnotCultureSpecificDayPostfixMetadata();
 
-        public void SetCulture(CultureInfo cultureInfo) => _currentCulture = cultureInfo;
 
-        public void Format(StringBuilder sb)
+        public void Format(StringBuilder sb, CultureInfo cultureInfo)
         {
-            sb.Insert(sb.Length, _currentCulture.DateTimeFormat.GetDayName(DateTime.Now.DayOfWeek));
+            sb.Insert(sb.Length, cultureInfo.DateTimeFormat.GetDayName(DateTime.Now.DayOfWeek));
         }
     }
 }